Integrating the new circular buffer in the respective files
