+++ include/hydra/CMakeLists.txt
@@ -27,6 +27,7 @@ install(FILES
   hydra_socket.h
   hydra_stats.h
   hydra_resource.h
+  hydra_resampler.h
   hydra_core.h
   hydra_server.h
   hydra_client.h
